Fix warning about cast to uint32 in net man

+++ Weave_Server/Weave_Server/src/ServerNetworkManager.cpp
@@ -178,7 +178,7 @@ void ServerNetworkManager::SendFeedMessagePacket( ClientProxyPtr aClient, const 
     OutputMemoryBitStream packet = {};
     packet.Write( FeedMessagePacket );
 
-    packet.WriteBits( aMsg, strlen( aMsg ) );
+    packet.WriteBits( aMsg, static_cast< UINT32 >( strlen( aMsg ) ) );
 
     SendPacket( packet, aClient->GetEndpoint() );
 }
